FIX: cambiado tipo parametro bookingId de int a long en search y delete booking

+++ src/main/java/business/booking/BookingAS.java
@@ -10,7 +10,7 @@ public interface BookingAS {
 
     public Result<BookingTOA> updateBooking(ModifyBookingRequestSOAP bookingSOAP);
 
-    public Result<Void> deleteBooking(int id);
+    public Result<Void> deleteBooking(long id);
 
-    public Result<BookingTOA> readBooking(int id);
+    public Result<BookingTOA> readBooking(long id);
 }
